Notifications table - renamed page.utility to pagination.utility
